+++ object.c
@@ -30,7 +30,7 @@ static Obj* allocateObject(size_t size, ObjType type) {
 
 // Allocate a new closure object and return its address.
 ObjClosure* newClosure(ObjFunction* function) {
-    ObjUpvalue** upvalues = ALLOCATE(ObjUpvalue*, (size_t)function->upvalueCount);
+    ObjUpvalue** upvalues = ALLOCATE(ObjUpvalue*, function->upvalueCount);
 
     for (int i = 0; i < function->upvalueCount; i++) {
         upvalues[i] = NULL;
@@ -104,7 +104,7 @@ ObjString* takeString(char* chars, int length) {
     ObjString* interned = tableFindString(&vm.strings, chars, length, hash);
 
     if (interned != NULL) {
-        FREE_ARRAY(char, chars, (size_t)length + 1);
+        FREE_ARRAY(char, chars, length + 1);
         return interned;
     }
 
@@ -119,7 +119,7 @@ ObjString* copyString(const char* chars, int length) {
 
     if (interned != NULL) return interned;
 
-    char* heapChars = ALLOCATE(char, (size_t)length+1);
+    char* heapChars = ALLOCATE(char, length+1);
     memcpy(heapChars, chars, length);
     heapChars[length] = '\0';
     return allocateString(heapChars, length, hash);
